tests: Bluetooth: Unicast client BSIM used group size as stream_cnt
Instead of setting the stream_cnt based on the number
of streams configurer, we use the number of streams actually
added to the group, and then configure that amount.

This should ensure that we do not try to configure any
streams that we have not already added to the group.

(cherry picked from commit bebb98efb1739de5edc29475b80be105d6f1f7de)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: bebb98efb1739de5edc29475b80be105d6f1f7de
Change-Id: I9025ecec07cc24913f4950852c7ea412f041d893
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4230385
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>