soc: posix: inf_clock: add C++ support to NATIVE_TASK
Explicitly define function pointer as (void*)() and not (void*).
In C this cast is done implicitly, but C++ does not allow it.

Also, the const is moved to the correct location. Now it's a constant
pointer to a function(void) that returns void, instead of a pointer
to a function(void) returning (const void).

(cherry picked from commit 3293d8057db701e525483394cd73403c513a3fab)

Original-Signed-off-by: Hein Wessels <heinwessels93@gmail.com>
GitOrigin-RevId: 3293d8057db701e525483394cd73403c513a3fab
Change-Id: I217f8819dd211371d8cdcf073b4e7c5b620b8087
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4197271
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>