boards: arm: mimimxrt685_evk_cm33: workaround to support sys_reboot
On the RT685, the sys_reboot API does not function correctly. This is
due to the fact that the external flash must be reset at power on. In
order to configure the ROM to reset the flash, a OTP shadow register
can be programmed with the GPIO port and pin of the flash reset pin.

Program these bits at boot when the CONFIG_REBOOT symbol is set,
so that sys_reboot works as expected.

(cherry picked from commit 86ae9512f13fa605910bbd68bf7846736c810407)

Original-Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: 86ae9512f13fa605910bbd68bf7846736c810407
Change-Id: Ib5ccc0fd22d0e577c2696f82c257ee7b1e389920
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4287547
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>