doc: update documentation for shell library change
The SHELL_SUBCMD_DICT_SET_CREATE macro was previous changed in PR #51231
to require a string comment as part of the macro but the documentation
was not updated to reflect that. This patch updates the documentation to
relect the change in the shell library.

(cherry picked from commit 57e6532b0a317111a3012c50ded77d4788b80058)

Original-Signed-off-by: Jason Yuan <jasonyuan@google.com>
GitOrigin-RevId: 57e6532b0a317111a3012c50ded77d4788b80058
Change-Id: Id98d36338645cc48cfd2b4ec04c991eb551d3a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4263768
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>