drivers: usb_dc_rpi_pico: avoid infinite unhandled irq retriggers
This driver enables a number of interrupts it does not attempt to handle.
This results in "unhandled IRQ: 0x...." messages being printed, and the
interrupt handler retriggers immediately again, and this happens again
and again forver, because nothing ends up clearing the interrupt.

This change implements very limited handling of these interrupts. A custom
warning is logged, and the interrupt is cleared.

This change does not imply that doing this is sufficient. More changes may
need to be implemented to more gracefully re-start transactions or
re-arm some endpoints, but this is one step in the right direction,
and at least the OS doesn't freeze up.

(cherry picked from commit 9a7f9ec0baee08eec73729d688a680095296658f)

Original-Signed-off-by: Purdea Andrei <andrei@purdea.ro>
GitOrigin-RevId: 9a7f9ec0baee08eec73729d688a680095296658f
Change-Id: I88ded949b943de1822582ab379259a2be4a54d92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4224526
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>