riscv: Do not remove ESF when SOC_ISR_SW_UNSTACKING
When CONFIG_SOC_ISR_SW_UNSTACKING is defined, it's up to the custom soc
code to remove the ESF, because the software-managed part of the ESF is
depending on the hardware. Fix this in the ISR code.

(cherry picked from commit 61a204c8310584f3e43229ad09eb0ea82ff39d93)

Original-Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: 61a204c8310584f3e43229ad09eb0ea82ff39d93
Change-Id: I9927d2cb3e300c51cf61afc241673abe9f258cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4212752
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>