doc: communication: update stale info
At this point I think it's clear that discord is more important than
email as a discussion forum for the project, so clarify that on the
communication and collaboration page.

Email is still important, though (e.g. Discord is currently not
available in China) -- and we are not deprecating it at all.

(cherry picked from commit 5c0a07fcf00ede3282be264e997c07000a8a892f)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 5c0a07fcf00ede3282be264e997c07000a8a892f
Change-Id: I499ed76f729c66fc168a40b2055e65fb458e268f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4222748
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>