native: Set stdout and stderr to line buffered
By default, stdout is line buffered when it is connected to a
terminal.
But when output is a pipe, it is block buffered.
This causes a problems when the test binary is integrated
into a test framework that reads the pipe.

(cherry picked from commit aadca692b3f1c53aad97f11febc1657ce695bd4b)

Original-Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>
GitOrigin-RevId: aadca692b3f1c53aad97f11febc1657ce695bd4b
Change-Id: I5bf1456aac38cd7c1598fd45ff4e150f697e9a02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4271572
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>