Bluetooth: Audio: Handle GATT/ISO race condition in unicast client
When releasing a stream causing the CIS to disconnect, there was
a race condition between the CIS disconnect event and the
idle state notification from the unicast server.

This ensures that if we are disconnecting a CIS as part of the
release, then we wait for the CIS disconnect event before calling
the "released" callback.

(cherry picked from commit ad304697abb0f5957f29876ed2204e361d0c3d38)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: ad304697abb0f5957f29876ed2204e361d0c3d38
Change-Id: Id09edaef1b707b8d77e1b7ab052d34b77d7af583
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4271569
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>