drivers: timer: stm32u5 lptimer waits for DIER complete
On the stm32U5, when modifying the DIER register of the LPTIM peripheral,
a new write operation to can only be performed when the previous write
operation is completed and before going-on.
This is done with a function call for better readability.

(cherry picked from commit df4fa7088da7818cc8db92e50626ae67578b8dea)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: df4fa7088da7818cc8db92e50626ae67578b8dea
Change-Id: I03f488a9f474a7e5d7be7c2c99759f29983cfc87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4241540
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>