drivers: regulator: common: skip voltage change at init if already valid
The current implementation always sets the voltage before enabling,
even if the current voltage is in the allowed range.

This has some side effects, i.e. for PMIC regulators that are
pre-programmed for a specific value but allow voltage changes during
runtime. The side effect being that the regulator will always be reset
to the lower value of the voltage range at init.

Another usecase would be when a bootloader sets a specific voltage then
loads an application that uses the same driver.

The proposed fix is to evaluate the current voltage and try to bring
the actual voltage in range if the current voltage is not valid
according to the min/max constraints.

Tested on custom SAMD20 board with a custom RK816 PMIC driver.

(cherry picked from commit 867c25480117e31ea2c510b414391ba9f43b9875)

Original-Signed-off-by: Ionut Catalin Pavel <iocapa@iocapa.com>
GitOrigin-RevId: 867c25480117e31ea2c510b414391ba9f43b9875
Change-Id: I5905f925d3b2c14243c43e5227b2d675c9ae7209
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4337862
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>