samples: mesh_badge: fix BT mesh provisioning error
The BT Mesh Configuration Client API is used in an asynchronous manner
and the app key was not added before attempt binding it to mesh models.
This caused the binding functions to fail and resulted in a fail when
trying to to send mesh messages.

This commit registers a callback that is called when the app key is
assigned. After this callback is called, the binding functions are
called.

(cherry picked from commit 098749bb4554b0ef8052d7d56d871280e0c70f65)

Original-Signed-off-by: Jonas Remmert <j.remmert@phytec.de>
GitOrigin-RevId: 098749bb4554b0ef8052d7d56d871280e0c70f65
Change-Id: I5c02cde850a2566ec5a42ed49bedcd579e0c699b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4277218
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>