Bluetooth: Controller: Fix ull_prepare_dequeue for skipped events
Fix ull_prepare_dequeue to not skip events when preempt
does not match the event in the head of the prepare queue.
The head of the prepare queue does not match when
ull_prepare_dequeue has put the head of the queue to the
last of the queue when it is calling lll_resume interface.
This happens when there is already an active event which
needs a preempt timeout to be setup and there is another
non-resume event in the pipeline which now becomes the
head. The fix ensure to restore the order of the events
if a preempt timeout was to be setup.

Currently ull_prepare_dequeue loops through all events
before stopping at the original way the queue was. This is
not efficient and is a scope for improvement in future.

(cherry picked from commit 0d54ca876168b7d5d28219af3f3f1a247da49209)

Original-Signed-off-by: Vinayak Kariappa Chettimada <vich@nordicsemi.no>
GitOrigin-RevId: 0d54ca876168b7d5d28219af3f3f1a247da49209
Change-Id: I799ef76dc0d74bd4da80d6f52257f03d1db51038
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4356686
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>