drivers: sensor: lsm6dso: round up to the nearest odr
It was too strict to require the caller to pass in an accurate odr as a
parameter. This patch is to round the odr up to the nearest one, when odr
does not match.

(cherry picked from commit 9c9304218b9ec9bbec6f7431a13b5646142a8dee)

Original-Signed-off-by: Zhang Lixu <lixu.zhang@intel.com>
GitOrigin-RevId: 9c9304218b9ec9bbec6f7431a13b5646142a8dee
Change-Id: I53325925657f80c996c98c24ca4e04c243e5fb40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4422801
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>