Bluetooth: Audio: CAP Initiator handle unexpected ASE changes
When performing the CAP Unicast Start procedure, the remote
server(s) may change the state of the ASE before the procedure
is complete.

If the state change is unexpected, we stop the procedure and
call the callback, and then the caller can decide what to
do after that (presumably call the stop procedure for the
streams that may be setup).

(cherry picked from commit 3ab925ecefaac28c66200fec27743be89bb06761)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 3ab925ecefaac28c66200fec27743be89bb06761
Change-Id: I5774ceca0ca2da0a8aa2ab8c6ffafa99a6fb7ea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4358088
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>