linker: lld: see if clang has a preferred linker
This asks the clang if it has its own preference for ld.lld.
This is to mirror what we are doing to find GNU ld, and to
make sure we are using the linker clang is using.

(cherry picked from commit aa8f81211875da4317899d83671daafdf0b66c24)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: aa8f81211875da4317899d83671daafdf0b66c24
Change-Id: I1fabceb81e55a1bb0801cd3b4fbb437f23e6a85d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4384636
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>