Samples: Bluetooth: Fix scanning restart for broadcast audio sink
If waiting on the first semaphore in the main loop times out, the main
loop starts from the beginning and attempts to start scanning for
broadcast sources. But since the first semaphore was not passed, the
scan_recv_cb callback was not called yet and thus the previously started
scanning was not stopped. As a result, an attempt to start the scanning
will fail with -EALREADY and the main thread will end. To solve this,
-EALREADY is now ignored when checking the return value of
bt_audio_broadcast_sink_scan_start.

(cherry picked from commit 5a488c389116d690842a5940174f791cbd7252cd)

Original-Signed-off-by: Ladislav Podivin <ladislav.podivin@tietoevry.com>
GitOrigin-RevId: 5a488c389116d690842a5940174f791cbd7252cd
Change-Id: I7eff984ffa2c7c6b854c4299d5c2c769123a04aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4303405
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>