Tests: bluetooth: controller: add correct number of max connections
The number of connections for the tx_buffer_alloc needs to be at least 3
For one of the tests it was set to default 1, which leads to unspecified
behaviour.

This is fixed by setting the maximum nr. of connections to 4

(cherry picked from commit a395093acbd731fcb6cdd7f98db999f68a22c5bb)

Original-Signed-off-by: Andries Kruithof <andries.kruithof@nordicsemi.no>
GitOrigin-RevId: a395093acbd731fcb6cdd7f98db999f68a22c5bb
Change-Id: I1652bc2e8a35de393241d56c9492e91d4b81212d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4303407
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>