Bluetooth: Host: Fix doc; bt_id_create does not accept pub addr
`bt_id_create` does not accept public addresses, but the phrasing that
is removed in this change can be interpreted as "replace the public
address with a different public address". This was confusing users, so
it will be removed.

The intended meaning of the original sentence may have been "use the
provided address for BT_ID_DEFAULT, instead of using the public
address". Amending the sentence to something like that was considered.
But, there was no consensus on the usefulness of this addition, so it
was left out for brevity.

(cherry picked from commit 30926ec0d360e8964172fffbf09bf57751a4afff)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 30926ec0d360e8964172fffbf09bf57751a4afff
Change-Id: I993eb2852749c732fa17f7be09dfb9ed3af15570
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4299172
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>