drivers: pwm: pwm_mcux_sctimer: duty cycle fix
This is a fix for a driver bug that assumes a
user will want a new pwm channel for a new
pwm signal if they decide to change the
period length of the pwm. In some cases,
this creates a noticable change of
duty cycle accuracy.

(cherry picked from commit 5081f6aa28edc818a47dda3868b169132496ca99)

Original-Signed-off-by: Emilio Benavente <emilio.benavente@nxp.com>
GitOrigin-RevId: 5081f6aa28edc818a47dda3868b169132496ca99
Change-Id: Ib9bece12e31bee625163125d369e9ff71ef6986f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4331684
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>