drivers/gna: No empty structs
Empty structs are incompatible when building with C++, which specifies
their size as 1 byte.  Technically they are illegal in C; GCC has
always supported them as an extension, but with a size of *zero*
bytes.

The upshot is that we can't have them in headers that may be presented
to a C++ compiler.  Add a placeholder field.

We only get the resulting compiler warning with -std=c++98 it seems
like, which is why this went undetected.

(cherry picked from commit cc46968c9719172c105f538ade6c53b950b42c74)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: cc46968c9719172c105f538ade6c53b950b42c74
Change-Id: Idd8783fc40f9451f9b3a7d83d5676dbe2629a731
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4303393
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>