Bluetooth: ATT: Introduce `bt_att_mtu`
This is a refactoring, only visible inside `att.c`.

Give the expression `chan->chan.tx.mtu` the name `bt_att_mtu`. Use it
when the intention is to get the ATT MTU property of a channel.

This is done in preparation a more complex expression for `bt_att_mtu`,
since the expression currently incorrect. The fix will come in a later
commit in the same PR.

(cherry picked from commit 314838969517c973051968ecf38610182fb8677c)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 314838969517c973051968ecf38610182fb8677c
Change-Id: Ia55ec73ccf06019ae2dff450c7db4d9b6dbea565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4474892
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>