Samples: zbus: benchmark: Exclude nrf52_bsim
For the POSIX arch, this test relays on the native_posix RTC.
But that does not exist in the nrf52_bsim.
And without something to measure actual execution time,
it does not make sense to benchmark in the POSIX arch.
So disable this test for this platform,
and add a clear build error to warn users.

(cherry picked from commit eed1c32d5177cca0b39997abcaa15c0172e29492)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: eed1c32d5177cca0b39997abcaa15c0172e29492
Change-Id: I15b5a630e95ebd32300d57a69ddcaa3775139a49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4315523
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>