samples/bluetooth/mtu_update: Relax RSSI requirement
The sample was only willing to connect if the pair RSSI level
was over -40dBm. That is ok when testing physically keeping
devices at less than ~1 meter from each other.
But in CI we test with a default attenuation of 60dB between
devices and the Tx power is 0dBm.
Which causes the tests based on this sample to fail.

Fix it.

(cherry picked from commit a8bd8e21f6669103ebbb0ddaa3a696068b24e818)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: a8bd8e21f6669103ebbb0ddaa3a696068b24e818
Change-Id: I501a98e3d343ac45fbcc54dfd73e709b3e9efcd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4315545
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>