samples: drivers: virtualization: add ivshmem doorbell sample
Provide a sample for ivshmem-doorbell running under virtualization
(QEMU). The sample relies on existing tools (ivshmem-server,
ivshmem-client) that must be run at the host (unix).

The sample can be run on userspace or kernelspace, and on arm64 (qemu,
kvm) and x86_64 (qemu) platforms.

This execution of this code is quite complex, thus the README should be
made available in the official documentation. This is why the code is a
sample and not a test.

(cherry picked from commit 1feaa9db066aad14a3f12e68255fc3a21eaa1195)

Original-Signed-off-by: Rodrigo Cataldo <rodrigo.cataldo@huawei.com>
Original-Co-authored-by: Henri Xavier <datacomos@huawei.com>
GitOrigin-RevId: 1feaa9db066aad14a3f12e68255fc3a21eaa1195
Change-Id: If3b9df8ead718b51b3384ab42484b291784bb6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4410671
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>