tests: drivers: build_all: add build for i2c_xilinx_axi
There is currently not an in-tree Zephyr board which enables the
i2c_xilinx_axi driver as part of its standard configuration. (It can be
enabled fairly easily with the Digilent Arty board, but that requires
modifying the ARM-provided FPGA block design in Vivado and rebuilding
the FPGA image, so it cannot be used with the default FPGA image.)

Add a build_all target to provide build coverage for this driver.

(cherry picked from commit c99900acfd5377f97b0e8f6650ad3861efdc3392)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: c99900acfd5377f97b0e8f6650ad3861efdc3392
Change-Id: If1e3185018aedda06b98c010895ef20eb8ab5dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4432947
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>